+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/ReactivePolarisCircuitBreakerFactory.java
@@ -17,13 +17,21 @@
 
 package com.tencent.cloud.polaris.circuitbreaker;
 
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 
 import com.tencent.cloud.polaris.circuitbreaker.common.PolarisCircuitBreakerConfigBuilder;
+import com.tencent.cloud.polaris.circuitbreaker.config.PolarisCircuitBreakerProperties;
 import com.tencent.cloud.polaris.circuitbreaker.util.PolarisCircuitBreakerUtils;
 import com.tencent.polaris.api.core.ConsumerAPI;
+import com.tencent.polaris.api.utils.ThreadPoolUtils;
 import com.tencent.polaris.circuitbreak.api.CircuitBreakAPI;
+import com.tencent.polaris.client.util.NamedThreadFactory;
 
+import org.springframework.beans.factory.DisposableBean;
 import org.springframework.cloud.client.circuitbreaker.ReactiveCircuitBreaker;
 import org.springframework.cloud.client.circuitbreaker.ReactiveCircuitBreakerFactory;
 
@@ -33,7 +41,7 @@ import org.springframework.cloud.client.circuitbreaker.ReactiveCircuitBreakerFac
  * @author seanyu 2023-02-27
  */
 public class ReactivePolarisCircuitBreakerFactory extends
-		ReactiveCircuitBreakerFactory<PolarisCircuitBreakerConfigBuilder.PolarisCircuitBreakerConfiguration, PolarisCircuitBreakerConfigBuilder> {
+		ReactiveCircuitBreakerFactory<PolarisCircuitBreakerConfigBuilder.PolarisCircuitBreakerConfiguration, PolarisCircuitBreakerConfigBuilder> implements DisposableBean {
 
 	private Function<String, PolarisCircuitBreakerConfigBuilder.PolarisCircuitBreakerConfiguration> defaultConfiguration =
 			id -> {
@@ -49,9 +57,19 @@ public class ReactivePolarisCircuitBreakerFactory extends
 
 	private final ConsumerAPI consumerAPI;
 
-	public ReactivePolarisCircuitBreakerFactory(CircuitBreakAPI circuitBreakAPI, ConsumerAPI consumerAPI) {
+	private final ScheduledExecutorService cleanupService = Executors.newSingleThreadScheduledExecutor(
+			new NamedThreadFactory("sct-reactive-circuitbreaker-cleanup", true));
+
+	public ReactivePolarisCircuitBreakerFactory(CircuitBreakAPI circuitBreakAPI, ConsumerAPI consumerAPI,
+			PolarisCircuitBreakerProperties polarisCircuitBreakerProperties) {
 		this.circuitBreakAPI = circuitBreakAPI;
 		this.consumerAPI = consumerAPI;
+		cleanupService.scheduleWithFixedDelay(
+				() -> {
+					getConfigurations().clear();
+				},
+				polarisCircuitBreakerProperties.getConfigurationCleanupInterval(),
+				polarisCircuitBreakerProperties.getConfigurationCleanupInterval(), TimeUnit.MILLISECONDS);
 	}
 
 	@Override
@@ -73,4 +91,8 @@ public class ReactivePolarisCircuitBreakerFactory extends
 		this.defaultConfiguration = defaultConfiguration;
 	}
 
+	@Override
+	public void destroy() {
+		ThreadPoolUtils.waitAndStopThreadPools(new ExecutorService[]{cleanupService});
+	}
 }