+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 - [feat:support TSF router.](https://github.com/Tencent/spring-cloud-tencent/pull/1403)
 - [feat:upgrade nearby router and add namespace nearby router.](https://github.com/Tencent/spring-cloud-tencent/pull/1404)
 - [fix:fix contract reporting bug when using TSF.](https://github.com/Tencent/spring-cloud-tencent/pull/1406)
+- [fix: fix PolarisCircuitBreakerConfiguration not clear when gateway invoke by wildcard apis](https://github.com/Tencent/spring-cloud-tencent/pull/1425)